Conflicts:
	Projects/Common/RubezhAPI/RubezhAPI.csproj
	Projects/FireAdministrator/Modules/GroupControllerModule/PumpStation/ViewModels/PumpStationViewModel.cs

+++ Projects/GKImitator/GKImitator/ViewModels/DescriptorViewModel.cs
@@ -18,14 +18,16 @@ namespace GKImitator.ViewModels
 		public BaseDescriptor GKBaseDescriptor { get; private set; }
 		public BaseDescriptor KauBaseDescriptor { get; private set; }
 		public GKBase GKBase { get { return GKBaseDescriptor.GKBase; } }
-		public int DescriptorNo { get; private set; }
+		public int GKDescriptorNo { get; private set; }
+		public int KauDescriptorNo { get; private set; }
 		public ushort TypeNo { get; private set; }
 		readonly List<ushort> AdditionalShortParameters;
 
 		public DescriptorViewModel(BaseDescriptor descriptor)
 		{
 			GKBaseDescriptor = descriptor;
-			DescriptorNo = descriptor.GetDescriptorNo();
+			GKDescriptorNo = descriptor.GKBase.GKDescriptorNo;
+			KauDescriptorNo = descriptor.GKBase.KAUDescriptorNo;
 
 			SetAutomaticRegimeCommand = new RelayCommand(OnSetAutomaticRegime);
 			SetManualRegimeCommand = new RelayCommand(OnSetManualRegime);
@@ -53,17 +55,17 @@ namespace GKImitator.ViewModels
 
 		void OnStateBitChanged(GKStateBit stateBit, bool isActive, ImitatorJournalItem additionalJournalItem = null)
 		{
+			ImitatorJournalItem journalItem = null;
 			if (isActive)
 			{
 				CurrentOnDelay = 0;
 				CurrentOffDelay = 0;
 				CurrentHoldDelay = 0;
 				TurningState = TurningState.None;
-				var journalItem = new ImitatorJournalItem();
 
 				if (stateBit == GKStateBit.On)
 				{
-					journalItem = additionalJournalItem ?? new ImitatorJournalItem(2, 9, 2, 0);
+					journalItem = new ImitatorJournalItem(2, 9, 2, 0);
 					if (HoldDelay != 0)
 					{
 						CurrentHoldDelay = HoldDelay;
@@ -73,7 +75,7 @@ namespace GKImitator.ViewModels
 
 				if (stateBit == GKStateBit.TurningOn)
 				{
-					journalItem = additionalJournalItem ?? new ImitatorJournalItem(2, 9, 4, 0);
+					journalItem = new ImitatorJournalItem(2, 9, 4, 0);
 					if (OnDelay != 0)
 					{
 						CurrentOnDelay = OnDelay;
@@ -83,7 +85,7 @@ namespace GKImitator.ViewModels
 
 				if (stateBit == GKStateBit.TurningOff)
 				{
-					journalItem = additionalJournalItem ?? new ImitatorJournalItem(2, 9, 5, 0);
+					journalItem = new ImitatorJournalItem(2, 9, 5, 0);
 					if (OffDelay != 0)
 					{
 						CurrentOffDelay = OffDelay;
@@ -95,7 +97,10 @@ namespace GKImitator.ViewModels
 
 				if (stateBit == GKStateBit.Off)
 				{
-					journalItem = additionalJournalItem ?? new ImitatorJournalItem(2, 9, 3, 3);
+					AdditionalShortParameters[0] = 0;
+					AdditionalShortParameters[1] = 0;
+					AdditionalShortParameters[2] = 0;
+					journalItem = new ImitatorJournalItem(2, 9, 3, 3);
 					SetStateBit(GKStateBit.Attention, false);
 					SetStateBit(GKStateBit.Fire1, false);
 					SetStateBit(GKStateBit.Fire2, false);
@@ -103,17 +108,17 @@ namespace GKImitator.ViewModels
 
 				if (stateBit == GKStateBit.Norm)
 				{
-					journalItem = additionalJournalItem ?? new ImitatorJournalItem(2, 14, 0, 0);
+					journalItem = new ImitatorJournalItem(2, 14, 0, 0);
 				}
 
 				if (stateBit == GKStateBit.Fire1)
 				{
-					journalItem = additionalJournalItem ?? new ImitatorJournalItem(2, 2, 0, 0);
+					journalItem = new ImitatorJournalItem(2, 2, 0, 0);
 				}
 
 				if (stateBit == GKStateBit.Fire2)
 				{
-					journalItem = additionalJournalItem ?? new ImitatorJournalItem(2, 3, 0, 0);
+					journalItem = new ImitatorJournalItem(2, 3, 0, 0);
 				}
 
 				if (stateBit == GKStateBit.Ignore)
@@ -126,9 +131,11 @@ namespace GKImitator.ViewModels
 					journalItem = new ImitatorJournalItem(2, 4, 0, 0);
 				}
 
-				AddJournalItem(journalItem);
-				RecalculateOutputLogic();
+				AddJournalItem(additionalJournalItem ?? journalItem);
 			}
+
+			if (additionalJournalItem != null || journalItem != null)
+				RecalculateOutputLogic();
 		}
 
 		void InitializeTypeNo()
@@ -245,41 +252,49 @@ namespace GKImitator.ViewModels
 		public int CurrentCardNo { get; set; }
 
 		public bool HasCard { get; private set; }
-
-		public List<byte> GetStateBytes(int no)
+		public List<byte> GetStateBytes(int no, DatabaseType databaseType)
 		{
-			var result = new List<byte>();
-
-			result.AddRange(ToBytes((short)TypeNo));
-
-			var controllerAddress = GKBaseDescriptor.ControllerAdress;
-			result.AddRange(ToBytes((short)controllerAddress));
-
-			var addressOnController = GKBaseDescriptor.AdressOnController;
-			result.AddRange(ToBytes((short)addressOnController));
-
-			var physicalAddress = GKBaseDescriptor.PhysicalAdress;
-			result.AddRange(ToBytes((short)physicalAddress));
-
-			result.AddRange(GKBaseDescriptor.Description);
-
-			var serialNo = 0;
-			result.AddRange(IntToBytes(serialNo));
-
-			result.AddRange(IntToBytes(StatesToInt()));
-
-			foreach (var additionalShortParameter in AdditionalShortParameters)
+			lock (locker)
 			{
-				result.AddRange(ShortToBytes(additionalShortParameter));
-			}
+				var result = new List<byte>();
 
-			if(HasCard)
-			{
-				result.RemoveRange(52, 4);
-				result.InsertRange(52, IntToBytes(CurrentCardNo));
-			}
+				result.AddRange(ToBytes((short)TypeNo));
 
-			return result;
+				if (databaseType == DatabaseType.Gk)
+				{
+					var controllerAddress = GKBaseDescriptor.ControllerAdress;
+					result.AddRange(ToBytes((short)controllerAddress));
+
+					var addressOnController = GKBaseDescriptor.AdressOnController;
+					result.AddRange(ToBytes((short)addressOnController));
+				}
+
+				var physicalAddress = GKBaseDescriptor.PhysicalAdress;
+				result.AddRange(ToBytes((short)physicalAddress));
+
+				if (databaseType == DatabaseType.Gk)
+				{
+					result.AddRange(GKBaseDescriptor.Description);
+				}
+
+				var serialNo = 0;
+				result.AddRange(IntToBytes(serialNo));
+
+				result.AddRange(IntToBytes(StatesToInt()));
+
+				foreach (var additionalShortParameter in AdditionalShortParameters)
+				{
+					result.AddRange(ShortToBytes(additionalShortParameter));
+				}
+				
+				if (databaseType == DatabaseType.Gk && HasCard)
+				{
+					result.RemoveRange(52, 4);
+					result.InsertRange(52, IntToBytes(CurrentCardNo));
+				}
+
+				return result;
+			}
 		}
 
 		List<byte> ToBytes(short shortValue)
@@ -308,7 +323,7 @@ namespace GKImitator.ViewModels
 				journalItem.ObjectDeviceType = (short)(GKBaseDescriptor.GKBase as GKDevice).Driver.DriverTypeNo;
 				journalItem.ObjectDeviceAddress = (short)(((GKBaseDescriptor.GKBase as GKDevice).ShleifNo - 1) * 256 + (GKBaseDescriptor.GKBase as GKDevice).IntAddress);
 			}
-			using(var dbService = new DbService())
+			using (var dbService = new DbService())
 			{
 				dbService.ImitatorJournalTranslator.Add(journalItem);
 			}
